+++ api/routers/__tests__/apiRouter.test.js
@@ -1,5 +1,5 @@
 const request = require("supertest");
-const server = require("../server.js");
+const server = require("../../server.js");
 
 describe("GET /api", () => {
   it("returns 200", () => {
