+++ api/routers/__tests__/listRouter.test.js
@@ -1,6 +1,6 @@
 const request = require("supertest");
-const server = require("../server.js");
-const db = require("../../data/dbConfig");
+const server = require("../../server.js");
+const db = require("../../../data/dbConfig");
 
 beforeAll(() => db.seed.run());
